completed a full Model Matching example
It was a bit of a trial to construct a Model Matching example.
The book Discrete Time Control Systems by Ogata seems to be the
only good example.  The version provided by Dr. Adel Ghandakly is
significantly different and does not appear to be identical.
Nonetheless this method could still be useful.