mis-configured Control Law example
The Control Law example with prediction was mis-configured.
The multiplication of two systems (A*B) is not the same
as series(A,B) in the Octave control toolbox.  Oddly, it
does seem to be the same as (B*A).  In any case, series()
should be used since it is more explicit.